feat:Netty源码修改可以调试代码

+++ transport/src/test/java/io/netty/channel/PendingWriteQueueTest.java
@@ -267,9 +267,9 @@ public class PendingWriteQueueTest {
         assertTrue(promise3.isDone());
         assertTrue(promise3.isSuccess());
         assertTrue(channel.finish());
-        assertEquals(1L, channel.readOutbound());
-        assertEquals(2L, channel.readOutbound());
-        assertEquals(3L, channel.readOutbound());
+        assertEquals((double) 1L, channel.readOutbound());
+        assertEquals((double) 2L, channel.readOutbound());
+        assertEquals((double) 3L, channel.readOutbound());
     }
 
     @Test
@@ -292,8 +292,8 @@ public class PendingWriteQueueTest {
         assertTrue(channel.finish());
         assertTrue(promise.isDone());
         assertTrue(promise.isSuccess());
-        assertEquals(1L, channel.readOutbound());
-        assertEquals(2L, channel.readOutbound());
+        assertEquals((double) 1L, channel.readOutbound());
+        assertEquals((double) 2L, channel.readOutbound());
     }
 
     @Test
@@ -362,8 +362,8 @@ public class PendingWriteQueueTest {
         assertTrue(promise2.isSuccess());
         assertTrue(channel.finish());
 
-        assertEquals(1L, channel.readOutbound());
-        assertEquals(2L, channel.readOutbound());
+        assertEquals((double) 1L, channel.readOutbound());
+        assertEquals((double) 2L, channel.readOutbound());
         assertNull(channel.readOutbound());
         assertNull(channel.readInbound());
     }
